requirements.txt
kafka-python>=1.3.2
sklearn>=0.18.1
pandas>=0.19.2